Major reshuffling of the internals for View. Got rid of the viewElement() method (it's still there for legacy code), now View's keep track of their node directly. This should be a bit faster. Also made event observation a bit more clean: teardown _must_ stop observing any events to prevent leaks.
Added support for OutlineView. This is very basic: the model objects need to have a property (children) that is an array of their children. If you set the leafKeyPath property on OutlineView, it will use that to determine whether the object is a leaf or not. Otherwise, it will query for the children and count them. I'd like to add a method to determine the key paths for a specify object. That way you could use different childrenKeyPaths based on the type of the object.