Added some documentation.
Added initial support for the validationError change type. This is intended to support deferred validation, which is essentially what you get when you send data to a server. There are some issues to work out for how ChangeNotifications with type validationError should propagate, but the hooks are in place.

Fixed a bug where change notifications wouldn't occur for nested arrays (see the testSubArray function in kvo-array-test.js). Basically, I added a version of infoForKey on Array that simply calls infoForKey for each element in the array. This establishes the parent link correctly. One snag is that I don't synthesise a KeyInfo structure, because no one ever uses it for Arrays. This _might_ cause problems later...

Added a field to FieldGroup to allow presenting error messages immediately. This is helpful if your form only has one field. It makes sense to display the bubble right away.