Some major changes that have been brewing for a while.
extensions/JSONRPC: added option to supply a target in the redirect. It would be nice to be able to specify window chrome characteristics.

foundation/net/Deferred: added a scope parameter for callback and failure methods. This reduces the need for binding the method (which reduces an unnecessary function call).

foundation/additions/function-additions: added Function.delay and Function.repeating. While this might be a bit more typing that the bindAndDelay method added to the Function prototype, it should be better, because it turns the delayed/repeating method into something that can be canceled and re-started (for repeating). Function.prototype.bindAndDelay may eventually be deprecated.

ui/dom/element-ie: fixed some nasty bugs introduced when I "streamlined" the code for getStyles. Oops.

ui/dom/element: added Element.contains which is a faster way of determining containment that walking up the node tree. I can't believe didn't know about DOM element methods compareDocumentPosition and contains, what rock have I been hiding under?

ui/app/Page: added support for global mouseEventListeners. This is used by overlays to dismiss themselves if the visitor clicks outside their DIV. This will not swallow the original click, but it will cause the overlay to hide. Also substantially reworked the focus tracking code. I'd still like to be able to detect when the browser window in IE loses focus, but that doesn't seem to work.

ui/fx/Animator: fixed the code for animating class transitions to determine what the pending classname is rather than using the old value (which is still stuck on the node). This means that multiple calls to the Animator that add/remove classnames during the same event cycle will actually work correctly.

ui/views/FormControl: moved a significant amount of code from TextField down to the form control base class. This cleans up TextField and makes that code available to other form controls, like Slider. The only TODO here is the __showingPlaceholder field from TextField leaked into the base class in FormControl#validate. I suppose I could re-implement this method on TextField and handle the placeholder there...

ui/views/View: changed the binding name for the action argument from `actionData` to `argument`. This better reflects the name used in Cocoa.

Also tweaked the animationOptions to use the reverse flag for animations that remove the class name rather than add it. This makes the animation options much more logical.

When determining the real value of the delegate, the View will no longer delete the temporary method until an actual value is returned. So if the delegate is wrong, this temporary method will always execute.

Completely changed the way system classes are restored to a node when View#setClass is called. Instead of looking for specific classes, the code merely looks for a specific prefix (coherent.Style.PREFIX which is currently 'ui-'). This fixes the bug with OutlineView level classes (e.g. ui-outline-level-4) getting removed when the node has a class binding.

ui/views/Overlay: added the property `clickOutsideToDismiss` which does exactly what it sounds like it does -- when the visitor clicks anywhere outside the overlay, the overlay is sent setVisible(false). Additionally, there are some new constants for position of overlays (used by Bubble and the MapView overlays).

ui/views/Slider: added an incrementValue binding. This permits developers to specify the valid values between minimum and maximum. Some redundant code went away, because it was consolidated in FormControl.

ui/views/TextField: added animation options for the placeholder, this isn't really perfect yet, because when you click in the field, the cursor adopts the current style of the placeholder. That needs to change, but I think it may be a limitation of the browser, which will need the fancy MobileMe label trick to do right.

Added property methods for value which account for the placeholder value stashed in the node.

Added accessoryViews and completion support. If a delegate is declared, its completionsForFieldWithText method will be called for each keypress to determine what values to show to the visitor. At the moment, an accessory view is hand crafted to display the completions, but the goal is to allow developers to specify a view template to hide and show (like a colour or date picker).

ui/views/Bubble: it's no longer necessary to have an arrow for a bubble. It's still nice, but for the accessoryView on a TextField it might not be desirable.

ui/views/theme.js: the prefix for system classes is changed from 'c-' to 'ui-'. Also I removed the whacky array of styles based on the change to View#setClass.