Added a very basic PopupButton implementation. Just enough to display an AnchoredBubble. Of course, you need to specify the anchored bubble in your VIEW.
Fixed a parsing error when parsing ISO8601 dates.
Better support setting allowsEmptySelection to false in CollectionView.
Added clickInsideToDismiss option to Overlay -- this makes creating pop-up menus a lot easier.
Also added delegate callbacks to Overlay (willShowOverlay and willHideOverlay).
Added active state to View. This toggles the ui-active class on the view node.
Also added a callDelegate method which handles the jiggity-pokety necessary to call a method on the delegate (which may not be present).
Added a parentViewController property to ViewController. This allows view controllers to look up the chain and invoke methods on their ancestors.